requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# OpenAI and AI Libraries
openai==1.54.0

# LangChain and RAG Components
langchain==0.3.7
langchain-openai==0.2.8
langchain-pinecone==0.2.0
langchain-community==0.3.6

# Vector Database
pinecone-client==5.0.1

# Document Processing
docx2txt==0.8

# Database - PostgreSQL
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35

# FastAPI and Web Framework
fastapi==0.115.4
uvicorn==0.32.0
websockets==13.1
pydantic==2.9.2
pydantic[email]==2.9.2

# HTTP Client
httpx==0.27.2

# Utility Libraries
python-multipart==0.0.12

